move stripe prompts out of shared file prompts file
